+++ src/components/LoginModal/style.js
@@ -10,18 +10,6 @@ export const LoginWrap = styled.div`
   max-width: 300px;
 `;
 
-export const LoginButton = styled.button`
-  margin-top: 1.25rem;
-  padding: 1.25rem;
-  padding-left: 2.5rem;
-  position: relative;
-  font-size: ${({ theme }) => theme.fontSize.regular};
-  font-weight: ${({ theme }) => theme.fontWeight.bold};
-  border: 1px solid ${({ theme }) => theme.color.gray};
-  border-radius: 12px;
-  color: ${props => props.color || 'black'};
-`;
-
 const Icon = css`
   position: absolute;
   left: 25px;
